+++ src/main/java/roomescape/controller/dto/ReservationRequestDto.java
@@ -1,27 +1,22 @@
-package roomescape.domain;
+package roomescape.controller.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
-public class ReservationResponseDto {
-    private final long id;
+public class ReservationRequestDto {
+
     private final String name;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private final LocalDate date;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
     private final LocalTime time;
 
-    public ReservationResponseDto(long id, String name, LocalDate date, LocalTime time) {
-        this.id = id;
+    public ReservationRequestDto(String name, LocalDate date, LocalTime time) {
         this.name = name;
         this.date = date;
         this.time = time;
-    }
-
-    public long getId() {
-        return id;
     }
 
     public String getName() {
